requirements.txt
django==1.8.3
redis==2.10.3
django-redis-sessions==0.5.0
tornado==4.2.1
tornado-redis==2.4.18
sockjs-tornado==1.0.1

psycopg2
